Bathymmétrie de la masse d'eau au format raster

Bathymmétrie de la masse d'eau au format raster

bathymetry_lake(sppolygon, pts_bathy, cellsize = 10)

Arguments

sppolygon

un objet sf correspondant un polygone de la masse d'eau

pts_bathy

une collection de points -objet sf- de dimension XYZ ou Z : profondeur

cellsize

taille du pixel par défaut 10 m

Value

Bathymmétrie de la masse d'eau au format raster ; plusieurs algorithmes pour la création de la bathymétrie sont testés.

IDW

Inverse Distance Weighting a pour principe que chaque point spatialisé de profondeur est un point relié autres points environnants, le poids de relation entre les points varie selon la distance entre les points (relation forte entre deux points proche et relation faible pour deux points distants).
TPRS

Thin Plate Regression Spline

Soap Film Smooth

Le choix se fait juste par l'écart le plus petit entre les pts_bathy et le modèle -en valeur absolue-. Pour cela on utilise la fonction diffrasterpts.

References

https://fishandwhistle.net/post/2019/bathymetry-lake-volume-estimation-using-r/

https://fromthebottomoftheheap.net/2016/03/27/soap-film-smoothers/

Examples

library(lakemetrics)
bathymetry_lake(lac,bathy)
#> [inverse distance weighted interpolation]
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `depth_raster_output = purrr::map(data, rasterizeptsbathy)`.
#> ℹ In group 1: `name = "GAM_Soap"`.
#> Caused by warning in `new_CppObject_xp()`:
#> ! GDAL Message 1: +init=epsg:XXXX syntax is deprecated. It might return a CRS with a non-EPSG compliant axis order.
#> Simple feature collection with 5141 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 686404.1 ymin: 6484050 xmax: 687144.1 ymax: 6484900
#> Projected CRS: RGF93 v1 / Lambert-93
#> First 10 features:
#>       value        X       Y                 geometry
#> 1  3.375144 686764.1 6484050 POINT (686764.1 6484050)
#> 2  3.640453 686774.1 6484050 POINT (686774.1 6484050)
#> 3  4.599929 686784.1 6484050 POINT (686784.1 6484050)
#> 4  4.569227 686794.1 6484050 POINT (686794.1 6484050)
#> 5  3.428421 686804.1 6484050 POINT (686804.1 6484050)
#> 6  3.389413 686814.1 6484050 POINT (686814.1 6484050)
#> 7  3.529544 686724.1 6484060 POINT (686724.1 6484060)
#> 8  4.998648 686734.1 6484060 POINT (686734.1 6484060)
#> 9  5.717455 686744.1 6484060 POINT (686744.1 6484060)
#> 10 6.586077 686754.1 6484060 POINT (686754.1 6484060)